# specs/007-crud-departamentos-frontend/contracts/departamentos.openapi.yaml
openapi: 3.0.3
info:
  title: API de Departamentos para Frontend
  version: 1.0.0
  description: Contrato utilizado por el frontend para consulta y gestion de departamentos
servers:
  - url: http://localhost:8080
paths:
  /api/v1/departamentos:
    get:
      summary: Listar departamentos
      operationId: listarDepartamentos
      tags: [Departamentos]
      responses:
        '200':
          description: Lista de departamentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DepartamentoResponse'
        '401':
          description: No autenticado
        '403':
          description: Rol insuficiente
    post:
      summary: Crear departamento
      operationId: crearDepartamento
      tags: [Departamentos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartamentoCreateRequest'
      responses:
        '201':
          description: Departamento creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartamentoResponse'
        '400':
          description: Error de validacion
        '401':
          description: No autenticado
        '403':
          description: Rol insuficiente
        '409':
          description: Clave duplicada

  /api/v1/departamentos/{clave}:
    parameters:
      - in: path
        name: clave
        required: true
        schema:
          type: string
          maxLength: 50
    get:
      summary: Obtener departamento por clave
      operationId: obtenerDepartamento
      tags: [Departamentos]
      responses:
        '200':
          description: Departamento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartamentoResponse'
        '401':
          description: No autenticado
        '403':
          description: Rol insuficiente
        '404':
          description: Departamento no encontrado
    put:
      summary: Actualizar departamento
      operationId: actualizarDepartamento
      tags: [Departamentos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartamentoUpdateRequest'
      responses:
        '200':
          description: Departamento actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartamentoResponse'
        '400':
          description: Error de validacion
        '401':
          description: No autenticado
        '403':
          description: Rol insuficiente
        '404':
          description: Departamento no encontrado
        '409':
          description: Conflicto de concurrencia
    delete:
      summary: Eliminar departamento
      operationId: eliminarDepartamento
      tags: [Departamentos]
      responses:
        '204':
          description: Departamento eliminado
        '401':
          description: No autenticado
        '403':
          description: Rol insuficiente
        '404':
          description: Departamento no encontrado
        '409':
          description: Departamento con empleados asociados

components:
  schemas:
    DepartamentoCreateRequest:
      type: object
      required: [clave, nombre]
      properties:
        clave:
          type: string
          maxLength: 50
          example: IT
        nombre:
          type: string
          maxLength: 50
          example: Tecnologia

    DepartamentoUpdateRequest:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
          maxLength: 50
          example: Tecnologia y Sistemas

    DepartamentoResponse:
      type: object
      required: [clave, nombre]
      properties:
        clave:
          type: string
          maxLength: 50
        nombre:
          type: string
          maxLength: 50